requirements.txt
streamlit
groq
pillow
python-dotenv
pandas